+++ CHANGELOG.md
@@ -2,4 +2,5 @@
 
 ## v1.0.33.a
 * Fixed create recipe may be polluted during the thread.
+* Added BOTH support for AutoIO